adjust tests following the refactor
