+++ tests/json/expected/artist_catalognum.json
@@ -1,0 +1,157 @@
+[
+  {
+    "album": "Cybernetic Punk Unit",
+    "album_id": "https://blvckplvgue.bandcamp.com/album/process-404-cybernetic-punk-unit",
+    "artist": "Process 404",
+    "artist_id": "https://blvckplvgue.bandcamp.com",
+    "tracks": [
+      {
+        "title": "Metallic Dominance",
+        "track_id": "https://blvckplvgue.bandcamp.com/track/metallic-dominance",
+        "release_track_id": null,
+        "artist": "Process 404",
+        "artist_id": "https://blvckplvgue.bandcamp.com",
+        "length": 393,
+        "index": 1,
+        "media": "Digital Media",
+        "medium": 1,
+        "medium_index": 1,
+        "medium_total": 4,
+        "artist_sort": null,
+        "disctitle": null,
+        "artist_credit": null,
+        "data_source": "bandcamp",
+        "data_url": "https://blvckplvgue.bandcamp.com/album/process-404-cybernetic-punk-unit",
+        "lyricist": null,
+        "composer": null,
+        "composer_sort": null,
+        "arranger": null,
+        "track_alt": null,
+        "work": null,
+        "mb_workid": null,
+        "work_disambig": null,
+        "bpm": null,
+        "initial_key": null,
+        "genre": null
+      },
+      {
+        "title": "Cybernetic Punk Unit",
+        "track_id": "https://blvckplvgue.bandcamp.com/track/cybernetic-punk-unit",
+        "release_track_id": null,
+        "artist": "Process 404",
+        "artist_id": "https://blvckplvgue.bandcamp.com",
+        "length": 514,
+        "index": 2,
+        "media": "Digital Media",
+        "medium": 1,
+        "medium_index": 2,
+        "medium_total": 4,
+        "artist_sort": null,
+        "disctitle": null,
+        "artist_credit": null,
+        "data_source": "bandcamp",
+        "data_url": "https://blvckplvgue.bandcamp.com/album/process-404-cybernetic-punk-unit",
+        "lyricist": null,
+        "composer": null,
+        "composer_sort": null,
+        "arranger": null,
+        "track_alt": null,
+        "work": null,
+        "mb_workid": null,
+        "work_disambig": null,
+        "bpm": null,
+        "initial_key": null,
+        "genre": null
+      },
+      {
+        "title": "Megacity Corporation",
+        "track_id": "https://blvckplvgue.bandcamp.com/track/megacity-corporation",
+        "release_track_id": null,
+        "artist": "Process 404",
+        "artist_id": "https://blvckplvgue.bandcamp.com",
+        "length": 436,
+        "index": 3,
+        "media": "Digital Media",
+        "medium": 1,
+        "medium_index": 3,
+        "medium_total": 4,
+        "artist_sort": null,
+        "disctitle": null,
+        "artist_credit": null,
+        "data_source": "bandcamp",
+        "data_url": "https://blvckplvgue.bandcamp.com/album/process-404-cybernetic-punk-unit",
+        "lyricist": null,
+        "composer": null,
+        "composer_sort": null,
+        "arranger": null,
+        "track_alt": null,
+        "work": null,
+        "mb_workid": null,
+        "work_disambig": null,
+        "bpm": null,
+        "initial_key": null,
+        "genre": null
+      },
+      {
+        "title": "Social Overdose",
+        "track_id": "https://blvckplvgue.bandcamp.com/track/social-overdose",
+        "release_track_id": null,
+        "artist": "Process 404",
+        "artist_id": "https://blvckplvgue.bandcamp.com",
+        "length": 263,
+        "index": 4,
+        "media": "Digital Media",
+        "medium": 1,
+        "medium_index": 4,
+        "medium_total": 4,
+        "artist_sort": null,
+        "disctitle": null,
+        "artist_credit": null,
+        "data_source": "bandcamp",
+        "data_url": "https://blvckplvgue.bandcamp.com/album/process-404-cybernetic-punk-unit",
+        "lyricist": null,
+        "composer": null,
+        "composer_sort": null,
+        "arranger": null,
+        "track_alt": null,
+        "work": null,
+        "mb_workid": null,
+        "work_disambig": null,
+        "bpm": null,
+        "initial_key": null,
+        "genre": null
+      }
+    ],
+    "asin": null,
+    "albumtype": "album",
+    "va": false,
+    "year": 2020,
+    "month": 6,
+    "day": 12,
+    "label": "Black/Plague",
+    "mediums": 1,
+    "artist_sort": null,
+    "releasegroup_id": null,
+    "catalognum": "",
+    "script": null,
+    "language": null,
+    "country": "FR",
+    "style": "electronic",
+    "genre": "industrial, techno",
+    "albumstatus": "Official",
+    "media": "Digital Media",
+    "albumdisambig": null,
+    "releasegroupdisambig": null,
+    "artist_credit": null,
+    "original_year": null,
+    "original_month": null,
+    "original_day": null,
+    "data_source": "bandcamp",
+    "data_url": "https://blvckplvgue.bandcamp.com/album/process-404-cybernetic-punk-unit",
+    "discogs_albumid": null,
+    "discogs_labelid": null,
+    "discogs_artistid": null,
+    "comments": "Future as always been a fantasy\n\nBlack/Plague wants to take you to the future. This distant fantasy that continues to inspire so many artists. Will technology bring us progress, or even more chaos?\nAntoine aka Process 404 wants you to take part. This near future where technology creates a constant conflict in our daily life. By combining elements and sonority of cyberpunk culture with Electronic, Process 404 seek to prove his vision of the future and why so many people call Techno « the sound of tomorrow». You can expect 4/4 tracks loaded with dystopian sounds and robotics ammo.\n---\nMastering: Axel Picodot",
+    "albumtypes": "album"
+  }
+]
